Story #12533: Factorisation
