+++ api/api-archive-search/archive-search-commons/src/main/java/fr/gouv/vitamui/archives/search/common/service/ArchiveUnitService.java
@@ -27,10 +27,10 @@
 
 package fr.gouv.vitamui.archives.search.common.service;
 
-import fr.gouv.vitamui.archives.search.common.dto.JsonPatchDto;
-import fr.gouv.vitamui.archives.search.common.dto.MultiJsonPatchDto;
 import fr.gouv.vitamui.archives.search.common.dto.OperationIdDto;
 import fr.gouv.vitamui.archives.search.common.dto.UpdateArchiveUnitDto;
+import fr.gouv.vitamui.commons.api.dtos.JsonPatchDto;
+import fr.gouv.vitamui.commons.api.dtos.MultiJsonPatchDto;
 
 import java.util.Set;
