+++ api/api-archive-search/archive-search-external/src/main/java/fr/gouv/vitamui/archives/search/external/server/service/ArchiveUnitServiceImpl.java
@@ -27,12 +27,12 @@
 
 package fr.gouv.vitamui.archives.search.external.server.service;
 
-import fr.gouv.vitamui.archives.search.common.dto.JsonPatchDto;
-import fr.gouv.vitamui.archives.search.common.dto.MultiJsonPatchDto;
 import fr.gouv.vitamui.archives.search.common.dto.OperationIdDto;
 import fr.gouv.vitamui.archives.search.common.dto.UpdateArchiveUnitDto;
 import fr.gouv.vitamui.archives.search.common.rest.ArchiveUnitClient;
 import fr.gouv.vitamui.archives.search.common.service.ArchiveUnitService;
+import fr.gouv.vitamui.commons.api.dtos.JsonPatchDto;
+import fr.gouv.vitamui.commons.api.dtos.MultiJsonPatchDto;
 import fr.gouv.vitamui.commons.api.exception.ApplicationServerException;
 import fr.gouv.vitamui.commons.rest.client.AbstractHttpContext;
 import fr.gouv.vitamui.commons.rest.client.ExternalHttpContext;
