+++ api/api-archive-search/archive-search-internal/src/main/java/fr/gouv/vitamui/archive/internal/server/config/ConverterConfig.java
@@ -44,6 +44,7 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @ComponentScan("fr.gouv.vitamui.archives.search.common.dto.converter")
+@ComponentScan("fr.gouv.vitamui.commons.api.converter")
 public class ConverterConfig {
 
     @Bean
