+++ api/api-collect/collect-commons/src/main/java/fr/gouv/vitamui/collect/common/service/ArchiveUnitService.java
@@ -27,10 +27,10 @@
 
 package fr.gouv.vitamui.collect.common.service;
 
-import fr.gouv.vitamui.collect.common.dto.JsonPatchDto;
-import fr.gouv.vitamui.collect.common.dto.MultiJsonPatchDto;
 import fr.gouv.vitamui.collect.common.dto.OperationIdDto;
 import fr.gouv.vitamui.collect.common.dto.UpdateArchiveUnitDto;
+import fr.gouv.vitamui.commons.api.dtos.JsonPatchDto;
+import fr.gouv.vitamui.commons.api.dtos.MultiJsonPatchDto;
 
 import java.util.Set;
