+++ api/api-collect/collect-external/src/main/java/fr/gouv/vitamui/collect/external/server/service/ArchiveUnitServiceImpl.java
@@ -27,12 +27,12 @@
 
 package fr.gouv.vitamui.collect.external.server.service;
 
-import fr.gouv.vitamui.collect.common.dto.JsonPatchDto;
-import fr.gouv.vitamui.collect.common.dto.MultiJsonPatchDto;
 import fr.gouv.vitamui.collect.common.dto.OperationIdDto;
 import fr.gouv.vitamui.collect.common.dto.UpdateArchiveUnitDto;
 import fr.gouv.vitamui.collect.common.rest.ArchiveUnitClient;
 import fr.gouv.vitamui.collect.common.service.ArchiveUnitService;
+import fr.gouv.vitamui.commons.api.dtos.JsonPatchDto;
+import fr.gouv.vitamui.commons.api.dtos.MultiJsonPatchDto;
 import fr.gouv.vitamui.commons.api.exception.ApplicationServerException;
 import fr.gouv.vitamui.commons.rest.client.AbstractHttpContext;
 import fr.gouv.vitamui.commons.rest.client.ExternalHttpContext;
