+++ api/api-collect/collect-internal/src/main/java/fr/gouv/vitamui/collect/internal/server/config/ConverterConfig.java
@@ -33,6 +33,7 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @ComponentScan("fr.gouv.vitamui.collect.common.dto.converter")
+@ComponentScan("fr.gouv.vitamui.commons.api.converter")
 public class ConverterConfig {
