+++ api/api-collect/collect-internal/src/main/java/fr/gouv/vitamui/collect/internal/server/rest/ArchiveUnitController.java
@@ -27,12 +27,12 @@
 
 package fr.gouv.vitamui.collect.internal.server.rest;
 
-import fr.gouv.vitamui.collect.common.dto.JsonPatchDto;
-import fr.gouv.vitamui.collect.common.dto.MultiJsonPatchDto;
 import fr.gouv.vitamui.collect.common.dto.OperationIdDto;
 import fr.gouv.vitamui.collect.common.dto.UpdateArchiveUnitDto;
-import fr.gouv.vitamui.collect.common.model.JsonPatch;
 import fr.gouv.vitamui.collect.common.service.ArchiveUnitService;
+import fr.gouv.vitamui.commons.api.dtos.JsonPatch;
+import fr.gouv.vitamui.commons.api.dtos.JsonPatchDto;
+import fr.gouv.vitamui.commons.api.dtos.MultiJsonPatchDto;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
