+++ api/api-collect/collect-internal/src/main/java/fr/gouv/vitamui/collect/internal/server/service/ArchiveUnitServiceImpl.java
@@ -35,17 +35,17 @@ import fr.gouv.vitam.common.exception.InvalidParseOperationException;
 import fr.gouv.vitam.common.exception.VitamClientException;
 import fr.gouv.vitam.common.json.JsonHandler;
 import fr.gouv.vitam.common.model.RequestResponseOK;
-import fr.gouv.vitamui.collect.common.dto.BulkCommandDto;
-import fr.gouv.vitamui.collect.common.dto.JsonPatchDto;
-import fr.gouv.vitamui.collect.common.dto.MultiJsonPatchDto;
 import fr.gouv.vitamui.collect.common.dto.OperationIdDto;
 import fr.gouv.vitamui.collect.common.dto.UpdateArchiveUnitDto;
-import fr.gouv.vitamui.collect.common.dto.converter.JsonPatchDtoToUpdateMultiQueryConverter;
 import fr.gouv.vitamui.collect.common.dto.converter.UpdateArchiveUnitDtoToUpdateMultiQueryConverter;
-import fr.gouv.vitamui.collect.common.dto.converter.UpdateMultiQueriesToBulkCommandDto;
 import fr.gouv.vitamui.collect.common.exception.ArchiveUnitUpdateException;
 import fr.gouv.vitamui.collect.common.model.OperationId;
 import fr.gouv.vitamui.collect.common.service.ArchiveUnitService;
+import fr.gouv.vitamui.commons.api.converter.JsonPatchDtoToUpdateMultiQueryConverter;
+import fr.gouv.vitamui.commons.api.converter.UpdateMultiQueriesToBulkCommandDto;
+import fr.gouv.vitamui.commons.api.dtos.BulkCommandDto;
+import fr.gouv.vitamui.commons.api.dtos.JsonPatchDto;
+import fr.gouv.vitamui.commons.api.dtos.MultiJsonPatchDto;
 import fr.gouv.vitamui.commons.api.logger.VitamUILogger;
 import fr.gouv.vitamui.commons.api.logger.VitamUILoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
