+++ commons/commons-api/src/main/java/fr/gouv/vitamui/commons/api/dtos/JsonPatchDto.java
@@ -25,15 +25,17 @@
  * accept its terms.
  */
 
-package fr.gouv.vitamui.archives.search.common.model;
+package fr.gouv.vitamui.commons.api.dtos;
 
 import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-import java.util.ArrayList;
-
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+import lombok.experimental.Accessors;
 
 @Data
-@EqualsAndHashCode(callSuper = true)
-public class JsonPatch extends ArrayList<PatchCommand> {
+@NoArgsConstructor
+@Accessors(chain = true)
+public class JsonPatchDto {
+    @NonNull private String id;
+    @NonNull private JsonPatch jsonPatch;
 }
