+++ commons/commons-api/src/main/java/fr/gouv/vitamui/commons/api/dtos/PatchCommand.java
@@ -25,28 +25,19 @@
  * accept its terms.
  */
 
-package fr.gouv.vitamui.archives.search.common.dto.converter;
+package fr.gouv.vitamui.commons.api.dtos;
 
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import fr.gouv.vitam.common.database.builder.request.multiple.UpdateMultiQuery;
-import fr.gouv.vitamui.archives.search.common.dto.BulkCommandDto;
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.stereotype.Service;
+import com.fasterxml.jackson.databind.JsonNode;
+import lombok.Data;
+import lombok.experimental.Accessors;
 
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 
-@Service
-public class UpdateMultiQueriesToBulkCommandDto implements Converter<Set<UpdateMultiQuery>, BulkCommandDto> {
-    @Override
-    public BulkCommandDto convert(Set<UpdateMultiQuery> source) {
-        final List<ObjectNode> finalUpdateMultiQueries =
-            source.stream()
-                .map(UpdateMultiQuery::getFinalUpdate)
-                .peek(objectNode -> objectNode.remove("$roots"))
-                .peek(objectNode -> objectNode.remove("$filter"))
-                .collect(Collectors.toList());
-        return new BulkCommandDto(finalUpdateMultiQueries);
-    }
+@Data
+@Accessors(chain = true)
+public class PatchCommand {
+    @NotNull private PatchOperation op;
+    @NotBlank private String path;
+    @NotNull private JsonNode value;
 }
